backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
motor>=3.1.1
pydantic>=2.0.0
pydantic-settings>=2.8.0
pymongo>=4.3.3
email-validator>=2.0.0
bcrypt>=4.0.1
PyJWT>=2.6.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
flake8==6.1.0
pytest-cov==4.1.0

# Email
requests>=2.28.1
python-postmark>=0.6.0

# LLM Integration
langchain>=0.3.20
langchain_core>=0.3.41
langchain_openai>=0.3.7
langgraph>=0.3.5
tiktoken>=0.9.0
openai>=1.65.4
python-dotenv>=1.0.1 